* `EventingWrapper` and `AsyncEventingWrapper` improvements and cleanups
